requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
albumentations==0.5.2
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
flake8==3.8.4
fsspec==0.8.5
future==0.18.2
google-auth==1.25.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
mccabe==0.6.1
multidict==5.1.0
mypy==0.800
mypy-extensions==0.4.3
networkx==2.5
numpy==1.20.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.2.1
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytorch-lightning==1.1.7
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.18.1
scipy==1.6.0
Shapely==1.7.1
six==1.15.0
snowballstemmer==2.1.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.2.1
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
yarl==1.6.3
